Fix bug in built-in benchmark fitness functions (#9)
* Fix bug in built-in benchmark fitness functions

The built-in GBO implementations of Sphere and Rosenbrock contained a bug: the vector of dependent subfunctions was initialized to a certain size and then extended with further numbers, leaving the subfunction `0` in the initialized part to always be included in the list of dependent subfunctions. This commit/PR fixes that initialization issue.

The impact of this issue was primarily visible in the evaluation count needed to reach the value-to-reach in those two fitness functions, which tended to be inflated due to this bug.

This commit also makes two other changes:
* It fixes another bug which caused (real-valued) conditional distributions to be discarded. This is done by wrapping a call to `initializeDistributions` in the real-valued `population.cpp` in an `if`-check.
* It introduces a BBO sphere function. Previously, only a GBO version was available of this benchmark function, and since the Rosenbrock function had two versions, and I needed a BBO version of sphere for testing, I thought it would be beneficial to include this, too.

* Undo different formatting

* Fix time condition issue

* Fix distribution multiplier int/double bug

+++ gomea/src/real_valued/linkage_model.cpp
@@ -253,7 +253,7 @@ void linkage_model_rv_t::initializeDistributions()
 		distributions.push_back( new normal_distribution_t(group) );
 }
 			
-int linkage_model_rv_t::getDistributionMultiplier( int element_index )
+double linkage_model_rv_t::getDistributionMultiplier( int element_index )
 {
 	return( distributions[element_index]->distribution_multiplier );
 }
